+++ app/service/embedding_service.py
